Change value type to decimal

+++ app/services/metric_values/contracts/persist.rb
@@ -3,7 +3,7 @@ module MetricValues
     class Persist < App::Contract
       params do
         required(:timestamp).filled(:date_time)
-        required(:value).filled(:float)
+        required(:value).filled(:decimal)
         required(:metric_id).filled(:string)
       end
     end
